django/requirements.txt
Django==4.0.6
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-oauth-toolkit==2.1.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.13.1
channels==3.0.4
ccxt==1.90.67
gunicorn==20.1.0
Pillow==5.0.0